Runs in forecast mode with GFS+NWM forcing.
Works mainly on a bare-metal configuration. Additional tweaks are
required in order to use SLURM so that pre-processing steps can happen
concurrently. Some additional optimizations might be needed to make the
hydrology input generation faster. The NationalWaterModel class uses
parallelization, so be aware of this if testing on a front-end node.