bugfix: multipolygon generation
This bug would appear on meshes that are represented by more than one
exterior polygon. On the previous version, the code would create and
return only the first exterior polygon. This provides a fix for that.
Note that this bug doesn't really affect any of the meshes we have
tested for SCHISM, which are all represented by a single exterior
polygon.